Refactor for testing
That latestDeployment.Value > DateTime.UtcNow.AddHours(-24) logic need to be unit tested.  Before writing the unit tests please refactor this code to be more testable.  The DateTime.UtcNow will need to be put into a class where "Now" can be mocked, and all the logic in Program will need to be put in a class with dependency injection so we can mock Connection and ConfigurationData.  This class could be called DevOpsManager.